+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- CSV headers adapts to the language provided through HTTP_ACCEPT_LANGUAGE header to facilitate import in Mapotempo-Web [#196](https://github.com/Mapotempo/optimizer-api/pull/196)
+
 ### Changed
 
 ### Removed